+++ src/it/clojure.test-fail/src/test/java/sample/FailingTest.java
@@ -1,0 +1,33 @@
+package sample;
+
+import org.junit.jupiter.api.Disabled;
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+/**
+ * Demonstrates surefire's behavior with JUnit, useful for comparing and contrasting
+ * behavior and console output with that of vivid:clojure-maven-plugin:clojure.test
+ */
+class FailingTest {
+
+    @Test
+    void fail() {
+        // JUnit is expected to mark this test as having failed.
+        assertTrue(false);
+    }
+
+    @Test
+    void error() {
+        // This exception is not expected by JUnit;
+        // JUnit is expected to mark this test as being in error.
+        throw new IllegalStateException();
+    }
+
+    @Test
+    @Disabled
+    void disabled() {
+        assertTrue(false);
+    }
+
+}