+++ src/it/clojure.test-pass/verify.groovy
@@ -1,0 +1,23 @@
+/*
+ * Copyright 2020 The vivid:clojure-maven-plugin Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License. You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License
+ * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+ * or implied. See the License for the specific language governing permissions and limitations under
+ * the License.
+ */
+
+def buildLog = new File(basedir, "build.log")
+assert buildLog.exists()
+assert buildLog.text.contains("[INFO] Running nested.ns-test")
+assert buildLog.text.contains("[INFO] Running clojure-test-api-sorezore-test")
+assert buildLog.text.contains("[INFO] Tests run: 13, Assertions: 26, Failures: 0, Errors: 0, Time elapsed:")
+assert buildLog.text.contains("[INFO] BUILD SUCCESS")
+
+def junitReportXml = new File (basedir, "target/clojure-test-reports/all-tests.xml")
+assert junitReportXml.exists()