+++ src/it/clojure.test-skip/src/test/java/sample/PassingTest.java
@@ -1,0 +1,14 @@
+package sample;
+
+import org.junit.jupiter.api.Test;
+
+/**
+ * Demonstrates surefire's behavior with JUnit, useful for comparing and contrasting
+ * behavior and console output with that of vivid:clojure-maven-plugin:clojure.test
+ */
+class PassingTest {
+
+    @Test
+    void pass() {}
+
+}